Move to Common/Integrations/Profiler and ditch the placeholder disposable
