+++ Common/Integrations/Profiler/IProfilerApi.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Pathoschild.Stardew.Common.Integrations.Profiler;
+
+public interface IProfilerApi
+{
+    public IDisposable RecordSection(string ModId, string EventType, string Details);
+}
