--- ContentPatcher/Experiment/IProfilerApi.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ContentPatcher.Experiment
-{
-    public interface IProfilerApi
-    {
-        public IDisposable RecordSection(string ModId, string EventType, string Details);
-    }
-}
